ADD: refactor findByObjectId()
