+++ westack/datasource/memorykvconnector.go
@@ -8,6 +8,7 @@ import (
 	"github.com/fredyk/westack-go/westack/memorykv"
 	"github.com/google/uuid"
 	"github.com/spf13/viper"
+	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/bson/primitive"
 )
 
@@ -85,6 +86,36 @@ func (connector *MemoryKVConnector) FindMany(collectionName string, lookups *wst
 		documents = bytes
 	}
 	return NewFixedMongoCursor(documents), nil
+}
+
+func (connector *MemoryKVConnector) findObjectById(collectionName string, _id interface{}, lookups *wst.A) (*wst.M, error) {
+	db := connector.db
+	bucket := db.GetBucket(collectionName)
+	var idAsString string
+	switch _id.(type) {
+	case string:
+		idAsString = _id.(string)
+	case primitive.ObjectID:
+		idAsString = _id.(primitive.ObjectID).Hex()
+	case uuid.UUID:
+		idAsString = _id.(uuid.UUID).String()
+	}
+	var document wst.M
+	allBytes, err := bucket.Get(idAsString)
+	if err != nil {
+		return nil, err
+	}
+	if len(allBytes) == 0 {
+		return nil, errors.New("document not found")
+	} else if len(allBytes) > 1 {
+		return nil, errors.New("multiple documents found")
+	} else {
+		err = bson.Unmarshal(allBytes[0], &document)
+		if err != nil {
+			return nil, err
+		}
+		return &document, nil
+	}
 }
 
 func (connector *MemoryKVConnector) Count(collectionName string, lookups *wst.A) (int64, error) {
